requirements.txt
pandas>=2.2.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.3.0
PyYAML>=6.0.1
pytest>=8.0.0
pytest-html>=4.1.0
python-dotenv>=1.0.0
